Update documentation and add new executable
The `README.md` file was updated to include new images and a download link for the executable. The previous images (`image.png` and `image-1.png`) were removed. The project description was slightly modified for clarity.

Two binary image files were deleted, and a new executable file (`Debug_Auto-Foldering.exe`) was added, containing a newly compiled version of the application. These changes enhance the usability and presentation of the Auto Foldering application.